backend/smartMoney_project/requirements.txt
asgiref>=3.2.3
attrs>=19.3.0
dj-database-url>=0.5.0
Django>=3.0.7
django-cors-headers>=3.4.0
django-filter>=2.3.0
django-heroku>=0.3.1
djangorestframework>=3.11.0
gunicorn>=20.0.4
importlib-metadata>=1.6.0
Markdown>=3.2.2
more-itertools>=8.3.0
packaging>=20.4
parse>=1.15.0
parse-type>=0.5.2
pluggy>=0.13.1
psycopg2>=2.9.1
py>=1.8.1
pyparsing>=2.4.7
pytest>=5.4.3
pytz>=2019.3
selenium>=3.141.0
six>=1.15.0
sqlparse>=0.3.1
urllib3>=1.25.9
wcwidth>=0.2.3
whitenoise>=5.0.1
drf-yasg>=1.20.0
zipp>=3.1.0
pytest-django>=4.4.0